Fix error in reading in Diesel_transport_S_trend.xlsx
This error was caused by a small change in column coercion in readxl 1.0.0.
The update makes columns with no data become filled with type logical NA. This
file, however, had many columns with numeric data after the 100th row, which is
where readxl stops looking for data with which to determine the type of data in
the column. These values were therefore converted to TRUE because the column
had been coerced into logical values, causing a loss of data.