Allow multiple package library locations
This removes the necessity to have all the packages required by CEDS in the local user R
library and instead accepts them from the system R library as well. (Currently on PIC it
is necessary to install the packages FAOSTAT, readxl (v1.0.0), stringr, and pbapply
locally, but all the others are available system-wide.)

Additionally this updates the PIC scripts to R version 3.3.3, the newest available.